Switch to float for Dendry algorithm

+++ external/dendry/include/controlfunction.h
@@ -12,7 +12,7 @@ public:
 	/// <summary>
 	/// Evaluate the function in (x, y)
 	/// </summary>
-	double evaluate(double x, double y) const
+	float evaluate(float x, float y) const
 	{
 		return static_cast<const Implementation*>(this)->EvaluateImpl(x, y);
 	}
@@ -21,7 +21,7 @@ public:
 	/// Check whether a point is inside the domain of the function
 	/// </summary>
 	/// <returns>True if the point is inside the domain</return>
-	bool insideDomain(double x, double y) const
+	bool insideDomain(float x, float y) const
 	{
 		return static_cast<const Implementation*>(this)->InsideDomainImpl(x, y);
 	}
@@ -30,7 +30,7 @@ public:
 	/// Return the distance between the point and the inside of the domain
 	/// </summary>
 	/// <returns>The distance between the point and the inside of the domain</return>
-	double distToDomain(double x, double y) const
+	float distToDomain(float x, float y) const
 	{
 		return static_cast<const Implementation*>(this)->DistToDomainImpl(x, y);
 	}
@@ -39,7 +39,7 @@ public:
 	/// Return the minimum value of the control function
 	/// </summary>
 	/// <returns>The minimum value of the control function</return>
-	double minimum() const
+	float minimum() const
 	{
 		return static_cast<const Implementation*>(this)->MinimumImpl();
 	}
@@ -48,7 +48,7 @@ public:
 	/// Return the maximum value of the control function
 	/// </summary>
 	/// <returns>The maximum value of the control function</return>
-	double maximum() const
+	float maximum() const
 	{
 		return static_cast<const Implementation*>(this)->MaximumImpl();
 	}
