+++ external/dendry/include/perlin.h
@@ -1,6 +1,6 @@
 #ifndef PERLIN_H
 #define PERLIN_H
 
-double Perlin(double x, double y);
+float Perlin(float x, float y);
 
 #endif // PERLIN_H
